requirements.txt
python-telegram-bot>=20.0
flask>=3.0.0
requests>=2.31.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
Pillow>=9.0.0
werkzeug>=3.0.0
flask-cors>=4.0.0
jinja2>=3.1.0
cryptography>=3.4.8